Move embedded certs to security/
